+++ examples/storybook/stories/Wrapper/Themes.js
@@ -1,16 +1,18 @@
 import React from 'react';
-import { RichContentEditorBox, Page } from '../Components/StoryParts';
+import { Page, Section, ContentState } from '../Components/StoryParts';
 import exapmleState from '../../../../e2e/tests/fixtures/themeing-info.json';
 import Palette from '../Components/Palette';
 import { wixPalettes } from '../palettesExample';
 import { SegmentedToggle, FormField, Layout } from 'wix-style-react';
-import EditorWrapper from '../Components/EditorWrapper';
-import sourcecode from '!!raw-loader!../Components/EditorWrapper';
+import ExampleApplication from '../Components/ExampleApplication';
 
 export default () => {
   return (
     <Page title="Theme Palette">
       <ThemeSelector />
+      <Section title="Content State">
+        <ContentState json={exapmleState} />
+      </Section>
     </Page>
   );
 };
@@ -45,16 +47,12 @@ class ThemeSelector extends React.Component {
             </SegmentedToggle>
           </FormField>
         </Layout>
-
         <Palette palette={wixPalettes[selected]} />
-
-        <RichContentEditorBox sourcecode={sourcecode}>
-          <EditorWrapper
-            key={selected}
-            contentState={exapmleState}
-            palette={wixPalettes[selected]}
-          />
-        </RichContentEditorBox>
+        <ExampleApplication
+          key={selected}
+          initialState={exapmleState}
+          palette={wixPalettes[selected]}
+        />
       </React.Fragment>
     );
   }