+++ packages/wrapper/web/src/FullscreenRenderer.tsx
@@ -1,8 +1,10 @@
 import React, { Component, Fragment, Suspense, Children, ReactElement } from 'react';
-import { RichContentProps } from './RichContentWrapperTypes';
+import { InitialState, RichContentProps } from './RichContentWrapperTypes';
 
 interface Props {
   children: ReactElement;
+  helpers?: RichContentProps['helpers'];
+  initialState: InitialState;
 }
 
 interface State {
@@ -10,24 +12,13 @@ interface State {
   index?: number;
   data?: any;
   Fullscreen?: any;
-  isMounted: boolean;
 }
 
 export default class FullscreenRenderer extends Component<Props, State> {
-  childProps: RichContentProps;
-
   constructor(props) {
     super(props);
     this.state = {
       isExpanded: false,
-      isMounted: false,
-    };
-    this.childProps = {
-      ...props.children.props,
-      helpers: {
-        ...props.children.props.helpers,
-        onExpand: this.onExpand,
-      },
     };
   }
 
@@ -35,7 +26,7 @@ export default class FullscreenRenderer extends Component<Props, State> {
     const Fullscreen = React.lazy(() =>
       import(/* webpackChunkName: "rce-ViewerModal"  */ './ViewerModal')
     );
-    this.setState({ Fullscreen, isMounted: true });
+    this.setState({ Fullscreen });
   }
 
   onExpand = (entityIndex, innerIndex = 0) =>
@@ -48,25 +39,26 @@ export default class FullscreenRenderer extends Component<Props, State> {
 
   setData = data => this.setState({ data });
 
+  addExpand = (helpers: object) => ({ ...helpers, onExpand: this.onExpand });
+
   render() {
-    const { isExpanded, index, data, Fullscreen, isMounted } = this.state;
-    const { children } = this.props;
+    const { isExpanded, index, data, Fullscreen } = this.state;
+    const { children, helpers: viewerHelpers = {}, initialState } = this.props;
+    const helpers = this.addExpand(viewerHelpers);
     return (
       <Fragment>
-        {Children.only(React.cloneElement(children, this.childProps))}
-        {isMounted && (
+        {Children.only(React.cloneElement(children, { helpers }))}
+        {Fullscreen && (
           <Suspense fallback={<div />}>
-            {Fullscreen && (
-              <Fullscreen
-                dataHook={'WrapperFullScreen'}
-                initialState={children.props.initialState || { entityMap: {} }}
-                isOpen={isExpanded}
-                images={data?.images || []}
-                onClose={this.onClose}
-                index={index}
-                setExpandModeData={this.setData}
-              />
-            )}
+            <Fullscreen
+              dataHook={'WrapperFullScreen'}
+              initialState={initialState || { entityMap: {} }}
+              isOpen={isExpanded}
+              images={data?.images || []}
+              onClose={this.onClose}
+              index={index}
+              setExpandModeData={this.setData}
+            />
           </Suspense>
         )}
       </Fragment>