+++ packages/wrapper/web/src/themeStrategy/themeTypes.d.ts
@@ -1,5 +1,22 @@
+interface PaletteColors {
+  actionColor: string;
+  bgColor: string;
+  textColor: string;
+  secondaryColor: string;
+  color7: string;
+  color4: string;
+}
+
+interface ThemeUtils {
+  fallbackColor: string;
+  fallbackColorBright: string;
+  isBright: (hexColor: string) => boolean;
+  adaptForeground: (actionColor: string) => string;
+  hexToRgbA: (hexColor: string, opacity: number) => string;
+}
+
 interface ThemeGeneratorFunction {
-  (colors: any, utils: any): any;
+  (colors: PaletteColors, utils: ThemeUtils): object;
 }
 
 interface ThemeProperties {